bp_runtime

Module extensions

Primitives that may be used for creating signed extensions for indirect runtimes.

Structs

GenericTransactionExtension

A simplified version of signed extensions meant for producing signed transactions and signed payloads in the client code.

GenericTransactionExtensionSchema

An implementation of TransactionExtensionSchema using generic params.

Traits

TransactionExtensionSchema

Trait that describes some properties of a TransactionExtension that are needed in order to send a transaction to the chain.

Type Aliases

BridgeRejectObsoleteHeadersAndMessages

The TransactionExtensionSchema for BridgeRejectObsoleteHeadersAndMessages.

ChargeTransactionPayment

The TransactionExtensionSchema for pallet_transaction_payment::ChargeTransactionPayment.

CheckEra

The TransactionExtensionSchema for frame_system::CheckEra.

CheckGenesis

The TransactionExtensionSchema for frame_system::CheckGenesis.

CheckNonZeroSender

The TransactionExtensionSchema for frame_system::CheckNonZeroSender.

CheckNonce

The TransactionExtensionSchema for frame_system::CheckNonce.

CheckSpecVersion

The TransactionExtensionSchema for frame_system::CheckSpecVersion.

CheckTxVersion

The TransactionExtensionSchema for frame_system::CheckTxVersion.

CheckWeight

The TransactionExtensionSchema for frame_system::CheckWeight.

PrevalidateAttests

The TransactionExtensionSchema for polkadot-runtime-common::PrevalidateAttests.

RefundBridgedParachainMessagesSchema

The TransactionExtensionSchema for RefundBridgedParachainMessages. This schema is dedicated for RefundBridgedParachainMessages signed extension as wildcard/placeholder, which relies on the scale encoding for () or ((), ()), or ((), (), ()) is the same. So runtime can contains any kind of tuple: (BridgeRefundBridgeHubRococoMessages) (BridgeRefundBridgeHubRococoMessages, BridgeRefundBridgeHubWestendMessages) (BridgeRefundParachainMessages1, ..., BridgeRefundParachainMessagesN)