Trait bp_runtime::extensions::TransactionExtensionSchema

pub trait TransactionExtensionSchema: Encode + Decode + Debug + Eq + Clone + StaticTypeInfo {
    type Payload: Encode + Decode + Debug + Eq + Clone + StaticTypeInfo;
    type Implicit: Encode + Decode + Debug + Eq + Clone + StaticTypeInfo;
}

Trait that describes some properties of a TransactionExtension that are needed in order to send a transaction to the chain.

Required Associated Types

type Payload: Encode + Decode + Debug + Eq + Clone + StaticTypeInfo

A type of the data encoded as part of the transaction.

type Implicit: Encode + Decode + Debug + Eq + Clone + StaticTypeInfo

Parameters which are part of the payload used to produce transaction signature, but don’t end up in the transaction itself (i.e. inherent part of the runtime).

Object Safety

This trait is not object safe.