Function bp_runtime::derive_account_id

pub fn derive_account_id<AccountId>(
    bridge_id: ChainId, 
    id: SourceAccount<AccountId>
) -> H256 where
    AccountId: Encode, 

Derive an account ID from a foreign account ID.

This function returns an encoded Blake2 hash. It is the responsibility of the caller to ensure this can be successfully decoded into an AccountId.

The bridge_id is used to provide extra entropy when producing account IDs. This helps prevent AccountId collisions between different bridges on a single target chain.

Note: If the same bridge_id is used across different chains (for example, if one source chain is bridged to multiple target chains), then all the derived accounts would be the same across the different chains. This could negatively impact users’ privacy across chains.