Module bp_runtime::extensions

Primitives that may be used for creating signed extensions for indirect runtimes.

Structs

GenericSignedExtension

A simplified version of signed extensions meant for producing signed transactions and signed payloads in the client code.

GenericSignedExtensionSchema

Traits

SignedExtensionSchema

Trait that describes some properties of a SignedExtension that are needed in order to send a transaction to the chain.

Type Aliases

BridgeRejectObsoleteHeadersAndMessages

The SignedExtensionSchema for BridgeRejectObsoleteHeadersAndMessages.

ChargeTransactionPayment

The SignedExtensionSchema for pallet_transaction_payment::ChargeTransactionPayment.

CheckEra

The SignedExtensionSchema for frame_system::CheckEra.

CheckGenesis

The SignedExtensionSchema for frame_system::CheckGenesis.

CheckNonZeroSender

The SignedExtensionSchema for frame_system::CheckNonZeroSender.

CheckNonce

The SignedExtensionSchema for frame_system::CheckNonce.

CheckSpecVersion

The SignedExtensionSchema for frame_system::CheckSpecVersion.

CheckTxVersion

The SignedExtensionSchema for frame_system::CheckTxVersion.

CheckWeight

The SignedExtensionSchema for frame_system::CheckWeight.

RefundBridgedParachainMessagesSchema

The SignedExtensionSchema for RefundBridgedParachainMessages. This schema is dedicated for RefundBridgedParachainMessages signed extension as wildcard/placeholder, which relies on the scale encoding for () or ((), ()), or ((), (), ()) is the same. So runtime can contains any kind of tuple: (BridgeRefundBridgeHubRococoMessages) (BridgeRefundBridgeHubRococoMessages, BridgeRefundBridgeHubWococoMessages) (BridgeRefundParachainMessages1, ..., BridgeRefundParachainMessagesN)