Trait bp_runtime::Chain

pub trait Chain: Send + Sync + 'static {
    type BlockNumber: Parameter + Member + MaybeSerializeDeserialize + Hash + Copy + Default + MaybeDisplay + AtLeast32BitUnsigned + FromStr + MaybeMallocSizeOf + AsPrimitive<usize>;
    type Hash: Parameter + Member + MaybeSerializeDeserialize + Hash + Ord + Copy + MaybeDisplay + Default + SimpleBitOps + AsRef<[u8]> + AsMut<[u8]> + MaybeMallocSizeOf;
    type Hasher: HashT<Output = Self::Hash>;
    type Header: Parameter + HeaderT<Number = Self::BlockNumber, Hash = Self::Hash> + MaybeSerializeDeserialize;
}

Minimal Substrate-based chain representation that may be used from no_std environment.

Associated Types

type BlockNumber: Parameter + Member + MaybeSerializeDeserialize + Hash + Copy + Default + MaybeDisplay + AtLeast32BitUnsigned + FromStr + MaybeMallocSizeOf + AsPrimitive<usize>

A type that fulfills the abstract idea of what a Substrate block number is.

type Hash: Parameter + Member + MaybeSerializeDeserialize + Hash + Ord + Copy + MaybeDisplay + Default + SimpleBitOps + AsRef<[u8]> + AsMut<[u8]> + MaybeMallocSizeOf

A type that fulfills the abstract idea of what a Substrate hash is.

type Hasher: HashT<Output = Self::Hash>

A type that fulfills the abstract idea of what a Substrate hasher (a type that produces hashes) is.

type Header: Parameter + HeaderT<Number = Self::BlockNumber, Hash = Self::Hash> + MaybeSerializeDeserialize

A type that fulfills the abstract idea of what a Substrate header is.

Loading content...