Trait bp_runtime::OwnedBridgeModule

pub trait OwnedBridgeModule<T: Config> {
    type OwnerStorage: StorageValue<T::AccountId, Query = Option<T::AccountId>>;
    type OperatingMode: OperatingMode;
    type OperatingModeStorage: StorageValue<Self::OperatingMode, Query = Self::OperatingMode>;

    const LOG_TARGET: &'static str;

    // Provided methods
    fn is_halted() -> bool { ... }
    fn ensure_owner_or_root(origin: T::RuntimeOrigin) -> Result<(), BadOrigin> { ... }
    fn ensure_not_halted() -> Result<(), OwnedBridgeModuleError> { ... }
    fn set_owner(
        origin: T::RuntimeOrigin,
        maybe_owner: Option<T::AccountId>
    ) -> DispatchResult { ... }
    fn set_operating_mode(
        origin: T::RuntimeOrigin,
        operating_mode: Self::OperatingMode
    ) -> DispatchResult { ... }
}

Bridge module that has owner and operating mode

Required Associated Types

type OwnerStorage: StorageValue<T::AccountId, Query = Option<T::AccountId>>

A storage entry that holds the module Owner account.

type OperatingMode: OperatingMode

Operating mode type of the pallet.

type OperatingModeStorage: StorageValue<Self::OperatingMode, Query = Self::OperatingMode>

A storage value that holds the pallet operating mode.

Required Associated Constants

const LOG_TARGET: &'static str

The target that will be used when publishing logs related to this module.

Provided Methods

fn is_halted() -> bool

Check if the module is halted.

fn ensure_owner_or_root(origin: T::RuntimeOrigin) -> Result<(), BadOrigin>

Ensure that the origin is either root, or PalletOwner.

fn ensure_not_halted() -> Result<(), OwnedBridgeModuleError>

Ensure that the module is not halted.

fn set_owner( origin: T::RuntimeOrigin, maybe_owner: Option<T::AccountId> ) -> DispatchResult

Change the owner of the module.

fn set_operating_mode( origin: T::RuntimeOrigin, operating_mode: Self::OperatingMode ) -> DispatchResult

Halt or resume all/some module operations.