Trait bp_runtime::StorageMapKeyProvider

pub trait StorageMapKeyProvider {
    type Hasher: StorageHasher;
    type Key: FullCodec;
    type Value: FullCodec;

    const MAP_NAME: &'static str;

    // Provided method
    fn final_key(pallet_prefix: &str, key: &Self::Key) -> StorageKey { ... }
}

Can be use to access the runtime storage key of a StorageMap.

Required Associated Types

type Hasher: StorageHasher

The same as StorageMap::Hasher1.

type Key: FullCodec

The same as StorageMap::Key1.

type Value: FullCodec

The same as StorageMap::Value.

Required Associated Constants

const MAP_NAME: &'static str

The name of the variable that holds the StorageMap.

Provided Methods

fn final_key(pallet_prefix: &str, key: &Self::Key) -> StorageKey

This is a copy of the frame_support::storage::generator::StorageMap::storage_map_final_key.

We’re using it because to call storage_map_final_key directly, we need access to the runtime and pallet instance, which (sometimes) is impossible.